// app/src/api/redmist/flags.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export enum Flags {
    /**
     * Value: 0
     */
    Unknown = 0,
    /**
     * Value: 1
     */
    Green = 1,
    /**
     * Value: 2
     */
    Yellow = 2,
    /**
     * Value: 3
     */
    Red = 3,
    /**
     * Value: 4
     */
    White = 4,
    /**
     * Value: 5
     */
    Checkered = 5,
    /**
     * Value: 6
     */
    Black = 6,
    /**
     * Purple 35mph.Value: 7
     */
    Purple35 = 7,
    /**
     * Purple 60kph. Value: 8
     */
    Purple60 = 8,
}